% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCNA.R
\name{collate_scna_segments}
\alias{collate_scna_segments}
\title{Title}
\usage{
collate_scna_segments(cohort, segmentation_files, as_grange = TRUE)
}
\arguments{
\item{as_grange}{}
}
\value{

}
\description{
Title
}
