% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_annotation.R
\name{collate_vcfs}
\alias{collate_vcfs}
\title{S3 Method for collating vcfs}
\usage{
collate_vcfs(vcf_list, tidy_function)
}
\arguments{
\item{tidy_function}{}
}
\value{

}
\description{
S3 Method for collating vcfs
}
