% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_annotation.R
\name{collate_vcfs.tn}
\alias{collate_vcfs.tn}
\title{Collate vcfs}
\usage{
\method{collate_vcfs}{tn}(vcf_list, tidy_function)
}
\arguments{
\item{tidy_function}{}
}
\value{

}
\description{
Collate vcfs
}
