% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_annotation2.R
\name{compare_t_cl}
\alias{compare_t_cl}
\title{Compare Variants in Matched Tumor and Cell Line Samples}
\usage{
compare_t_cl(cobrinik_vars, caller)
}
\arguments{
\item{caller}{}
}
\value{

}
\description{
Compare Variants in Matched Tumor and Cell Line Samples
}
