% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCNA.R
\name{create_seg_granges}
\alias{create_seg_granges}
\title{Create Segmentation Granges}
\usage{
create_seg_granges(karyo_seg)
}
\arguments{
\item{karyo_seg}{}
}
\value{

}
\description{
Create Segmentation Granges
}
