% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_annotation2.R
\name{divide_t_cl}
\alias{divide_t_cl}
\title{Split Variant Set into Tumor and Cell Line Sets}
\usage{
divide_t_cl(mydf)
}
\arguments{
\item{mydf}{}
}
\description{
Split Variant Set into Tumor and Cell Line Sets
}
