% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_annotation2.R
\name{filt_vcf_by_gold}
\alias{filt_vcf_by_gold}
\title{Filter VCFs by gold standard variants}
\usage{
filt_vcf_by_gold(vcf, gold_vars)
}
\arguments{
\item{gold_vars}{}
}
\description{
Filter VCFs by gold standard variants
}
