% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_annotation2.R
\name{filter_variant_set}
\alias{filter_variant_set}
\title{Fitler Set of Variants}
\usage{
filter_variant_set(my_variants, datatype, gnomad_threshold = 5e-04)
}
\arguments{
\item{datatype}{}
}
\description{
Fitler Set of Variants
}
