% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_annotation2.R
\name{five_author_intersect}
\alias{five_author_intersect}
\title{Compare variants with four prior studies}
\usage{
five_author_intersect(cobrinik_vars, reference_vars, reference_genes,
  caller)
}
\arguments{
\item{caller}{}
}
\description{
Compare variants with four prior studies
}
