% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCNA.R
\name{format_SCNA_data}
\alias{format_SCNA_data}
\title{Title}
\usage{
format_SCNA_data(dna_copy_object, samples, remove_pattern, sample_type = NULL)
}
\arguments{
\item{sample_type}{}
}
\value{

}
\description{
Title
}
