% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCNA.R
\name{load_seg_objs}
\alias{load_seg_objs}
\title{Load Segmentation Objects}
\usage{
load_seg_objs(segmentation_files)
}
\arguments{
\item{segmentation_files}{}
}
\value{

}
\description{
Load Segmentation Objects
}
