% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_waterfall_plot}
\alias{make_waterfall_plot}
\title{Make a waterfall plot}
\usage{
make_waterfall_plot(
  my_variants,
  mainPalette = "ggsci::default_igv",
  clindata,
  subset_genes = NULL,
  sample_order = NULL,
  clinLegCol = 5,
  mainLabelSize = 12,
  ...
)
}
\arguments{
\item{clindata}{}
}
\value{

}
\description{
Make a waterfall plot
}
