% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_annotation.R
\name{mutect2_pon_tidy}
\alias{mutect2_pon_tidy}
\title{Tidy panel of normals VCF from Mutect2}
\usage{
mutect2_pon_tidy(my_vcf)
}
\arguments{
\item{my_vcf}{}
}
\description{
Tidy panel of normals VCF from Mutect2
}
