% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCNA.R
\name{plot_base_seg}
\alias{plot_base_seg}
\title{Plot Base Segmentation}
\usage{
plot_base_seg(cov_granges, tn, kp, chr_select, cex, tr.i, tr.o, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Plot Base Segmentation
}
