% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOH.R
\name{plot_chrom_to_file}
\alias{plot_chrom_to_file}
\title{Plot Chromosome to File}
\usage{
plot_chrom_to_file(raw_cov_list, file_name, chr_select, ...)
}
\arguments{
\item{...}{}
}
\description{
Plot Chromosome to File
}
