% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOH.R
\name{plot_loh_granges}
\alias{plot_loh_granges}
\title{Title}
\usage{
plot_loh_granges(
  raw_cov_list,
  chr_select = "auto",
  mid.regs,
  marker_granges,
  cex = 0.6,
  marker_labels = "Peak Sites",
  group = FALSE,
  group_space = 0.1,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
