% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOH.R
\name{plot_loh_granges_to_file}
\alias{plot_loh_granges_to_file}
\title{Title}
\usage{
plot_loh_granges_to_file(raw_cov_list, file_name, chr_select = "auto",
  mid.regs, marker_granges, plotheight = 10, plotwidth = 50, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
