% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOH.R
\name{plot_loh_group}
\alias{plot_loh_group}
\title{Plot SCNA group}
\usage{
plot_loh_group(
  cov_granges,
  num_baf_granges,
  group_space = 0.1,
  kp,
  tr.i,
  tr.o,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Plot SCNA group
}
