% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCNA.R
\name{plot_scna_granges}
\alias{plot_scna_granges}
\title{Plot All Segmentation to Heatmap}
\usage{
plot_scna_granges(
  raw_cov_list,
  chr_select = "auto",
  marker_granges = NULL,
  marker_labels = "Peak Sites",
  cex = 0.6,
  group = FALSE,
  group_space = 0.1,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Plot All Segmentation to Heatmap
}
