% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCNA.R
\name{plot_scna_granges_to_file}
\alias{plot_scna_granges_to_file}
\title{Plot All Segmentation to Heatmap}
\usage{
plot_scna_granges_to_file(raw_cov_list, file_name, chr_select = "auto",
  marker_granges = NULL, file_ext = "pdf", plotwidth = 40,
  plotheight = 10, res = 400, ...)
}
\description{
Plot All Segmentation to Heatmap
}
