% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCNA.R
\name{plot_scna_group}
\alias{plot_scna_group}
\title{Plot SCNA group}
\usage{
plot_scna_group(raw_cov_list, num_seg_granges, group_space = 0.1, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Plot SCNA group
}
