% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_annotation2.R
\name{pon_subtract_m2_reynolds}
\alias{pon_subtract_m2_reynolds}
\title{Antijoin Mutect2 Reynolds Samples with Panel of Normals}
\usage{
pon_subtract_m2_reynolds(my_variants, my_pon)
}
\arguments{
\item{my_pon}{}
}
\description{
Antijoin Mutect2 Reynolds Samples with Panel of Normals
}
