% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_annotation2.R
\name{pon_subtract_m2_vars}
\alias{pon_subtract_m2_vars}
\title{Antijoin Mutect Variants with Panel of Normals}
\usage{
pon_subtract_m2_vars(my_variants, my_pon)
}
\arguments{
\item{my_pon}{}
}
\description{
Antijoin Mutect Variants with Panel of Normals
}
