% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webgestalt.R
\name{prep_stachelek_clindata}
\alias{prep_stachelek_clindata}
\title{prep stachelek clindata}
\usage{
prep_stachelek_clindata(clindata)
}
\arguments{
\item{clindata}{}
}
\value{

}
\description{
prep stachelek clindata
}
