% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_annotation2.R
\name{print_tl}
\alias{print_tl}
\title{print length of table (to show how many variant present)}
\usage{
print_tl(variant_df, tname)
}
\arguments{
\item{tname}{}
}
\description{
print length of table (to show how many variant present)
}
