% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_annotation2.R
\name{recalculate_geo}
\alias{recalculate_geo}
\title{Recalculate geo from webgestalt}
\usage{
recalculate_geo(geo_output, gene_input, intgenelength = 245)
}
\arguments{
\item{intgenelength}{}
}
\value{

}
\description{
Recalculate geo from webgestalt
}
