% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webgestalt.R
\name{run_webgestaltr}
\alias{run_webgestaltr}
\title{Run WebGestaltR}
\usage{
run_webgestaltr(
  gene_list,
  gene_list_path,
  enrichDatabase = "geneontology_Biological_Process",
  enrichMethod = "ORA",
  topThr = 25
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Run WebGestaltR
}
\examples{

}
