% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_annotation2.R
\name{split_m2_somatic_vars}
\alias{split_m2_somatic_vars}
\title{split mutect2 variant list into sample type dataframe list}
\usage{
split_m2_somatic_vars(m2_somatic_vars, m2_filt_variant_paths = NULL)
}
\arguments{
\item{m2_filt_variant_paths}{}
}
\description{
split mutect2 variant list into sample type dataframe list
}
