% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_annotation.R
\name{standardize_vcf_cols}
\alias{standardize_vcf_cols}
\title{Standardize Vcf column names for PON samples}
\usage{
standardize_vcf_cols(vcf_df)
}
\arguments{
\item{vcf_df}{}
}
\value{

}
\description{
Standardize Vcf column names for PON samples
}
