% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_waterfall_plot}
\alias{make_waterfall_plot}
\title{Make a Waterfall Plot}
\usage{
make_waterfall_plot(my_variants, clindata, gene_lab_size = 10)
}
\arguments{
\item{gene_lab_size}{}
}
\description{
Make a Waterfall Plot
}
